+++ ft_memset.c
@@ -6,11 +6,13 @@
 /*   By: hoysong <hoysong@student.42gyeongsan.      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/26 19:30:11 by hoysong           #+#    #+#             */
-/*   Updated: 2024/02/29 22:14:08 by hoysong          ###   ########.fr       */
+/*   Updated: 2024/03/22 11:55:36 by hoysong          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memset(void *dest, int n_src, unsigned long len)
+#include "libft.h"
+
+void	*ft_memset(void *dest, int n_src, size_t len)
 {
 	unsigned char	*ptr;
